Split HikingMetrics.jsx
No DB/API integration yet, but both components (InputHikingExercise.jsx and ViewHikingExercise.jsx) can now safely be invoked anywhere.
